Improve usage of lloovia_yaml. Now it only exports two functions.
Previously, to convert problems or solutions, first a Converter object
needed to be created and then the function to convert could be called.
In addition, if the Converter object was used more than once, the result
was wrong. The new structure is cleaner and less error-prone for users
of the code.